Deleted massive amounts of obsolete code
Some frontend functions had to be commented out for now, as they will be obsoleted soon

+++ utils/api_gen.py
@@ -14,7 +14,7 @@ from .utils import assert_dict_data_type
 
 def create_object_view(
     Serializer: serializers.ModelSerializer,
-    editable_attr_types: List[str],
+    editable_attrs: List[str],
     owner_path: str,
     owner_type: Literal['USER', 'PROFILE']
 ):
@@ -22,7 +22,7 @@ def create_object_view(
     Creates an API view for creating a given model
 
     `Serializer`: Model's serializer (also gives information on the Model)
-    `editable_attr_types`: Dictionary mapping the model's editable attributes to their types
+    `editable_attrs`: Dictionary mapping the model's editable attributes to their types
         ({'title': str, 'num': int})
     `owner_path`: Path/attribute to get the objects owner
         ('user', 'profile.user', 'deck__user')
@@ -34,7 +34,7 @@ def create_object_view(
     def view(request: WSGIRequest):
         if resp := assert_request_data_type(request, {'create_values': dict}):
             return resp
-        if msg := assert_dict_data_type(request.data['create_values'], editable_attr_types):
+        if msg := assert_dict_data_type(request.data['create_values'], editable_attrs):
             return Response({'message': msg}, status=400)
 
         model = Model.objects.create(
@@ -63,7 +63,12 @@ def get_object_view(
     @api_view(['GET'])
     @permission_classes([IsAuthenticated] if owner_path else [])
     def view(request: WSGIRequest, obj_id: int):
-        model, got = get_obj_or_404(Model, obj_id, request.user, owner_path)
+        model, got = get_obj_or_404(
+            Model=Model,
+            obj_id=obj_id,
+            user=request.user,
+            owner_path=owner_path,
+        )
         if not got:
             return model
 
@@ -100,14 +105,15 @@ def list_object_view(
 
 def edit_object_view(
     Serializer: serializers.ModelSerializer,
-    editable_attrs: List[str],
+    editable_attrs: dict,
     owner_path: Union[str, None],
 ):
     """
     Creates an API view for editing a given model
 
     `Serializer`: Model's serializer (also gives information on the Model)
-    `editable_attrs`: Which attributes you can edit on the Model, for security
+    `editable_attrs`: Dictionary mapping the model's editable attributes to their types
+        ({'title': str, 'num': int})
     `owner_path`: Path/attribute to get the objects owner
         ('user', 'profile.user', 'deck__user')
     """
@@ -124,7 +130,10 @@ def edit_object_view(
             return model
 
         for attr, value in request.data['edited_values'].items():
-            if attr not in editable_attrs:
+            if (
+                attr not in editable_attrs.keys() or
+                not isinstance(attr, editable_attrs[attr])
+            ):
                 return Response(
                     {'message': f'Cannot edit attribute `{attr}`'},
                     status=400,
@@ -169,11 +178,16 @@ def generate_base_api(
     app_name: str,
     model_name: str,
     Serializer: serializers.ModelSerializer,
-    editable_attr_types: dict,
+    editable_attrs: dict,
     owner_path: Union[str, None],
     owner_type: Literal['USER', 'PROFILE'],
     /,
     exclude_app_name: bool = False,
+    exclude_create: bool = False,
+    exclude_get: bool = False,
+    exclude_list: bool = False,
+    exclude_edit: bool = False,
+    exclude_delete: bool = False,
 ):
     """
     Generate a list of API paths for a model
@@ -181,39 +195,58 @@ def generate_base_api(
     `app_name`: Name of the app that houses the model
     `model_name`: Model's lowercase name
     `Serializer`: Model's serializer (also gives information on the Model)
-    `editable_attr_types`: Dictionary mapping the model's editable attributes to their types
+    `editable_attrs`: Dictionary mapping the model's editable attributes to their types
         ({'title': str, 'num': int})
     `owner_path`: Path/attribute to get the objects owner
         ('user', 'profile.user', 'deck__user')
     `exclude_app_name`: Exclude the app name from the generated URLs
+    `exclude_create ... exclude_delete`: Don't add a view for that function
     """
     base_name = f'{app_name}/{model_name}' if not exclude_app_name else f'{model_name}'
-    return [
-        # Create
-        path(f'{base_name}/create/', create_object_view(
-            Serializer=Serializer,
-            editable_attr_types=editable_attr_types,
-            owner_path=owner_path,
-            owner_type=owner_type,
-        )),
-        # Read
-        path(f'{base_name}/<int:obj_id>/', get_object_view(
-            Serializer=Serializer,
-            owner_path=owner_path,
-        )),
-        path(f'{base_name}/', list_object_view(
-            Serializer=Serializer,
-            owner_path=owner_path,
-        )),
-        # Update
-        path(f'{base_name}/<int:obj_id>/edit/', edit_object_view(
-            Serializer=Serializer,
-            editable_attrs=editable_attr_types.keys(),
-            owner_path=owner_path,
-        )),
-        # Delete
-        path(f'{base_name}/<int:obj_id>/delete/', delete_object_view(
-            Serializer=Serializer,
-            owner_path=owner_path,
-        )),
-    ]
+    views = []
+
+    if not exclude_create:
+        views.append(
+            path(f'{base_name}/create/', create_object_view(
+                Serializer=Serializer,
+                editable_attrs=editable_attrs,
+                owner_path=owner_path,
+                owner_type=owner_type,
+            ))
+        )
+
+    if not exclude_get:
+        views.append(
+            path(f'{base_name}/<int:obj_id>/', get_object_view(
+                Serializer=Serializer,
+                owner_path=owner_path,
+            ))
+        )
+
+    if not exclude_list:
+        views.append(
+            path(f'{base_name}/', list_object_view(
+                Serializer=Serializer,
+                owner_path=owner_path,
+            ))
+        )
+
+    if not exclude_edit:
+        views.append(
+            path(f'{base_name}/<int:obj_id>/edit/', edit_object_view(
+                Serializer=Serializer,
+                editable_attrs=editable_attrs.keys(),
+                owner_path=owner_path,
+            ))
+        )
+
+    # TODO: add rearrange view
+
+    if not exclude_delete:
+        views.append(
+            path(f'{base_name}/<int:obj_id>/delete/', delete_object_view(
+                Serializer=Serializer,
+                owner_path=owner_path,
+            ))
+        )
+    return views
